abstract tokenize_vllm and do not use token prompts on initial turn
