+++ verifiers/utils/response_utils.py
@@ -1,5 +1,9 @@
-from typing import cast
+from typing import Optional, cast
 
+from openai import AsyncOpenAI, BaseModel
+from openai.types.chat import ChatCompletionToolParam
+
+import verifiers as vf
 from verifiers.types import (
     ChatCompletion,
     ChatMessage,
@@ -125,3 +129,34 @@ async def parse_response_messages(
             response_text = response.choices[0].text or ""
         completion_messages = str(response_text)
     return completion_messages
+
+
+async def tokenize_vllm(
+    tokens_client: AsyncOpenAI,
+    messages: Messages,
+    tools: list[ChatCompletionToolParam] | None,
+    model: str,
+    default_body: dict = {},
+) -> list[int]:
+    """Tokenize messages using the vLLM /tokenize API."""
+    body = dict(
+        model=model,
+        messages=messages,
+        tools=tools,
+        **default_body,
+    )
+
+    # Copy from vllm/entrypoints/openai/protocol.py
+    class TokenizeResponse(BaseModel):
+        count: int
+        max_model_len: int
+        tokens: list[int]
+        token_strs: Optional[list[str]] = None
+
+    try:
+        tokenize_response = await tokens_client.post(
+            "/tokenize", body=body, cast_to=TokenizeResponse
+        )
+        return tokenize_response.tokens
+    except Exception as e:
+        raise vf.ModelError(e)
